COVID-0002: Adding .env | separated routes file
