+++ .env
@@ -1,0 +1,2 @@
+BASE_URL=http://localhost:8080/build
+API_REST=http://ec2-3-21-99-253.us-east-2.compute.amazonaws.com/api/v1/
